+++ src/main/java/fr/ul/acl/escape/monde/Heros.java
@@ -4,7 +4,7 @@ import fr.ul.acl.escape.monde.exceptions.MouvementNullException;
 
 public class Heros extends Personnage {
 
-    public Heros(float x, float y, int hauteur, int largeur, float vitesse) {
+    public Heros(double x, double y, double hauteur, double largeur, double vitesse) {
         super(x, y, hauteur, largeur, vitesse);
     }
